Organizacion correcta entre domain y ui

+++ app/src/main/java/com/example/mvp/domain/model/Training.kt
@@ -1,8 +1,7 @@
 package com.example.mvp.domain.model
-import com.example.mvp.ui.screens.training.TrainingType
 
 data class Training(
-    val id: Long = 0L,
+    val id: Int = 0,
     val name: String,
     val dateText: String,
     val durationMin: Int,
